+++ rest_framework/pagination.py
@@ -1,4 +1,6 @@
 from rest_framework import serializers
+from rest_framework.settings import api_settings
+from rest_framework.templatetags.rest_framework import replace_query_param
 
 # TODO: Support URLconf kwarg-style paging
 
@@ -7,30 +9,42 @@ class NextPageField(serializers.Field):
     """
     Field that returns a link to the next page in paginated results.
     """
+    page_field = 'page'
+
     def to_native(self, value):
         if not value.has_next():
             return None
         page = value.next_page_number()
         request = self.context.get('request')
-        relative_url = '?page=%d' % page
+        url = ''
         if request:
-            return request.build_absolute_uri(relative_url)
-        return relative_url
+            if api_settings.USE_ABSOLUTE_URLS:
+                url = request.build_absolute_uri()
+            else:
+                url = request.get_full_path()
+
+        return replace_query_param(url, self.page_field, page)
 
 
 class PreviousPageField(serializers.Field):
     """
     Field that returns a link to the previous page in paginated results.
     """
+    page_field = 'page'
+
     def to_native(self, value):
         if not value.has_previous():
             return None
         page = value.previous_page_number()
         request = self.context.get('request')
-        relative_url = '?page=%d' % page
+        url = ''
         if request:
-            return request.build_absolute_uri('?page=%d' % page)
-        return relative_url
+            if api_settings.USE_ABSOLUTE_URLS:
+                url = request.build_absolute_uri()
+            else:
+                url = request.get_full_path()
+
+        return replace_query_param(url, self.page_field, page)
 
 
 class PaginationSerializerOptions(serializers.SerializerOptions):